Day 4 slot C sub-slice 4C.d.viii: LankaEvents.Infrastructure unblock
LankaEvents.Infrastructure: 184 -> 0 ✅.

Same-shape mechanical sweeps applied per prior sub-slice patterns:
- `LankaConnect.SPLIT_PER_ENTITY[.*]` -> `LankaConnect.Infrastructure.Data[.*]`
  (placeholder namespace fix — same as 4C.d.vii applied to Infrastructure).
- `LankaConnect.BuildingBlocks.Domain.Shared[.*]` using deletion.
- SharedKernel.Money using added to files that reference Money/Currency.

Converter fixes (Money-operator-form pattern per Consult #13 Q2):
- MoneyConverter.cs: `Enum.TryParse<Currency>` -> `Currency.TryFromCode`;
  `Money.Create(...).Value` -> `new Money(...)`.
- NonNullableMoneyConverter.cs: same treatment.

Repository/service fixes:
- EventRepository.cs: `(decimal)` cast on Coordinates.Latitude/Longitude
  MapFrom (double -> decimal per stub GeoCoordinate signature).
- SeatHoldCleanupService.cs: fully-qualified `Domain.Common.IUnitOfWork`
  rewritten to `LankaConnect.BuildingBlocks.Domain.IUnitOfWork`.

Build state:
- Pre-4C.d.viii: 200 errors (LankaEvents.Infrastructure 184 dominated).
- Post-4C.d.viii: 929 errors. LankaEvents.Infrastructure 184 -> 0 ✅.
  * MASSIVE unmask cascade: LankaEvents.Api 1642 errors surfaced
    (same-shape but LankaEvents.Api is the biggest Api project consuming
    all LankaEvents handlers; Money/Currency signature changes cascaded
    through hundreds of controller endpoints).
  * Communications.Infrastructure 62 unchanged.
  * Payments.Infrastructure 18 unchanged.

Day 4 gate blocked on LankaEvents.Api 1642 + Communications/Payments Infra
+ TestUtilities. All same-shape mechanical per the pattern established
across 4C.d.iv through 4C.d.viii.

T-triggers: T6 (Money-signature migration extends into Infrastructure
converters + BackgroundService DI resolution).
S-class: n/a.

+++ src/Products/LankaEvents/LankaEvents.Infrastructure/Repositories/EventRepository.cs
@@ -559,8 +559,8 @@ public class EventRepository : ProductRepositoryBase<Event>, IEventRepository
                 Distance = _geoLocationService.CalculateDistanceKm(
                     latitude,
                     longitude,
-                    e.Location!.Coordinates!.Latitude,
-                    e.Location.Coordinates.Longitude
+                    (decimal)e.Location!.Coordinates!.Latitude,
+                    (decimal)e.Location.Coordinates.Longitude
                 )
             })
             .OrderBy(x => x.Distance)
